+++ azure-monitor-opentelemetry/samples/logging/correlated_logs.py
@@ -1,0 +1,32 @@
+# -------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License. See License in the project root for
+# license information.
+# --------------------------------------------------------------------------
+
+from logging import WARNING, getLogger
+
+from azure.monitor.opentelemetry import configure_azure_monitor
+from opentelemetry import trace
+
+configure_azure_monitor(
+    connection_string="<your-connection-string>",
+    logger_name=__name__,
+    logging_level=WARNING,
+    disable_metrics=True,
+    disable_tracing=True,
+)
+
+logger = getLogger(__name__)
+tracer = trace.get_tracer(__name__)
+
+logger.info("Uncorrelated info log")
+logger.warning("Uncorrelated warning log")
+logger.error("Uncorrelated error log")
+
+with tracer.start_as_current_span("Span for correlated logs"):
+    logger.info("Correlated info log")
+    logger.warning("Correlated warning log")
+    logger.error("Correlated error log")
+
+input()